Refactor the Office
